parse params

+++ src/utils.rs
@@ -339,10 +339,10 @@ pub fn bytes_to_blr(b: &[u8]) -> (Vec<u8>, Vec<u8>) {
     (blr, v)
 }
 
-pub fn convert_date(year: u32, month: u32, day: u32) -> [u8; 4] {
+pub fn convert_date(year: i32, month: u32, day: u32) -> [u8; 4] {
     // Convert date to BLR format data
     let i = month + 9;
-    let jy = year + (i / 12) - 1;
+    let jy = year as u32 + (i / 12) - 1;
     let jm = i % 12;
     let c = jy / 100;
     let j = (146097 * c) / 4 + (1461 * (jy - 100 * c)) / 4 + (153 * jm + 2) / 5 + day - 678882;
